refactor: 게시글 도메인에 게시글 사이즈 필드 추가

+++ src/main/java/com/openmpy/taleswiki/article/presentation/response/ArticleReadResponse.java
@@ -7,7 +7,7 @@ public record ArticleReadResponse(
         String title,
         String nickname,
         String content,
-        long version,
+        int version,
         LocalDateTime createdAt
 ) {
